requirements.txt
jupyter-book
jupyter
nbformat
nbclient
jupytext
myst-nb
jupyter-client
ipykernel
PyYAML

